implement gradle and fix error in dummy data

+++ app/src/main/java/com/refinaldy/newbibit/dummydata/Reksadana.kt
@@ -2,7 +2,7 @@ package com.refinaldy.newbibit.dummydata
 
 data class Reksadana(
     var investmentManager: String = "",
-    var oneYearReturn: Int = 0,
+    var oneYearReturn: Double = 0.00,
     var aum: String = "",
     var expenseratio: String = ""
